move files around, refactor configuration classes
