--- EZVIDIA/include/configurations/structs/GlobalConfiguration.hpp
@@ -1,12 +1,0 @@
-#pragma once
-#include<vector>
-#include<stdint.h>
-#include<string>
-#include"DisplayConfiguration.hpp"
-
-class GlobalConfiguration {
-public:
-	std::wstring name;
-	std::vector<DisplayConfiguration> displays;
-	uint32_t primaryGroup;
-};
