+++ EZVIDIA/src/drivers/DriverException.hpp
