+++ EZVIDIA/src/drivers/NvapiPathInfoWrapper.hpp
