+++ EZVIDIA/src/networking/NetworkException.hpp
