+++ EZVIDIA/src/utils/StringUtils.cpp
@@ -1,4 +1,4 @@
-#include "..\..\include\utils\StringUtils.hpp"
+#include "StringUtils.hpp"
 #define WIN32_LEAN_AND_MEAN
 #include <Windows.h>
